+++ synapse/config/__init__.py
@@ -1,0 +1,5 @@
+"""Synapse configuration package.
+
+Protocol Version: 1.0
+"""
+PROTOCOL_VERSION: str = "1.0"